requirements.txt
asgiref==3.6.0
click==8.1.3
Django==4.2.1
django-extensions==3.2.1
djangorestframework==3.14.0
pydotplus==2.0.2
pyparsing==3.0.9
pytz==2023.3
sqlparse==0.4.4
typer==0.9.0
typing_extensions==4.5.0
